Moved all presentation layer stuff to KatanaHook
`hook.py` now defines a KatanaHook class which can be subclassed to hook
into the output of Katana. This allows katana to be run
programmatically. The DefaultKatanaHook will output to the `.json`. The
JinjaKatanaHook will output to `.html` and `.json`, and the
`LoggingKatanaHook` will output to `.html`, `.json`, and log to the
console (preserving old functionality).