Fixed some weird completion errors.
python queue.Queue has no way to clear the contents of the queue which
causes long hangs when the threads are signaled to close but are in the
middle of interacting with the queue. To solve this, I nastily hacked
"clear" method together for the queue. It happens when setting the
`completed` property of katana. __If the stdlib changes, this will break.__